+++ app/services/ai/chatgpt_3_5_turbo/chatgpt_3_5_turbo.go
@@ -4,17 +4,16 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"strings"
+	"net/http"
+	"os"
+	"time"
 
-	cloudtasks "cloud.google.com/go/cloudtasks/apiv2"
 	"github.com/YukiTsuchida/conversational-ai-sns-bot/app/config"
 	"github.com/YukiTsuchida/conversational-ai-sns-bot/app/ent/conversations"
 	ai_model "github.com/YukiTsuchida/conversational-ai-sns-bot/app/models/ai"
 	"github.com/YukiTsuchida/conversational-ai-sns-bot/app/models/conversation"
 	"github.com/YukiTsuchida/conversational-ai-sns-bot/app/services/ai"
-	"google.golang.org/api/option"
-	taskspb "google.golang.org/genproto/googleapis/cloud/tasks/v2"
-	"google.golang.org/grpc"
+	"github.com/go-resty/resty/v2"
 
 	"github.com/YukiTsuchida/conversational-ai-sns-bot/app/ent"
 	"github.com/YukiTsuchida/conversational-ai-sns-bot/app/ent/chatgpt35turboconversationlog"
@@ -24,7 +23,7 @@ import (
 
 const modelName = "gpt-3.5-turbo"
 const tokenLimit = 3600 // 実際は4096だが、回答も含めて4096なので少なめにしておく
-const temperature = "1.0"
+const temperature = 1.0
 
 var _ ai.Service = (*aiServiceChatGPT3_5TurboImpl)(nil)
 
@@ -33,84 +32,67 @@ type aiServiceChatGPT3_5TurboImpl struct {
 }
 
 // CloudTasksに積むhttpリクエストのbody
-type Request struct {
-	CallBackUrl string    `json:"call_back_url"`
-	AIModel     string    `json:"ai_model"`
-	Temperature string    `json:"temperature"`
-	Messages    []Message `json:"messages"`
+// type Request struct {
+// 	CallBackUrl string    `json:"call_back_url"`
+// 	AIModel     string    `json:"ai_model"`
+// 	Temperature string    `json:"temperature"`
+// 	Messages    []Message `json:"messages"`
+// }
+
+// ChatGPT APIに投げるリクエストのbody型
+// Ref: https://platform.openai.com/docs/api-reference/chat/create
+type ChatGPTAPIRequest struct {
+	Model       string              `json:"model"`
+	Temperature float64             `json:"temperature"`
+	Messages    []ChatGPTAPIMessage `json:"messages"`
 }
 
-type Message struct {
+type ChatGPTAPIMessage struct {
 	Role    string `json:"role"`
-	Message string `json:"message"`
+	Content string `json:"content"`
 }
 
-func (ai *aiServiceChatGPT3_5TurboImpl) SendRequest(ctx context.Context, conversationID *conversation.ID) error {
+type ChatGPTAPIResponse struct {
+	ID      string `json:"id"`
+	Object  string `json:"object"`
+	Created int    `json:"created"`
+	Choices []struct {
+		Index   int `json:"index"`
+		Message struct {
+			Role    string `json:"role"`
+			Content string `json:"content"`
+		} `json:"message"`
+		FinishReason string `json:"finish_reason"`
+	} `json:"choices"`
+}
+
+func (ai *aiServiceChatGPT3_5TurboImpl) SendRequest(ctx context.Context, conversationID *conversation.ID) (*ai_model.Response, error) {
 	conversationIDInt, err := conversationID.ToInt()
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	logs, err := ai.db.Chatgpt35TurboConversationLog.Query().Where(chatgpt35turboconversationlog.HasConversationWith(conversations.IDEQ(conversationIDInt))).Order(ent.Asc(chatgpt35turboconversationlog.FieldCreatedAt)).All(ctx)
 	if err != nil {
-		return err
-	}
-
-	cloudTasksHost := config.CLOUDTASKS_HOST()     // local環境ではエミュレータを使うので環境変数からhostを指定する
-	cloudtasksParent := config.CLOUDTASKS_PARENT() // 「projects/%s/locations/%s」の部分
-	cloudtasksChild := "/queues/" + conversationID.ToString()
-
-	var client *cloudtasks.Client
-	var clientOpt option.ClientOption = nil
-	if cloudTasksHost != "" {
-		conn, _ := grpc.Dial(cloudTasksHost+":8123", grpc.WithInsecure())
-		clientOpt = option.WithGRPCConn(conn)
-		client, err = cloudtasks.NewClient(ctx, clientOpt)
-		if err != nil {
-			return err
-		}
-	} else {
-		client, err = cloudtasks.NewClient(ctx)
-		if err != nil {
-			return err
-		}
-	}
-
-	queuePath := cloudtasksParent + cloudtasksChild
-	_, err = client.GetQueue(ctx, &taskspb.GetQueueRequest{Name: queuePath})
-	if err != nil {
-		if strings.Contains(err.Error(), "NotFound") {
-			// queueが存在しなければ作る
-			createQueueRequest := taskspb.CreateQueueRequest{
-				Parent: cloudtasksParent,
-				Queue:  &taskspb.Queue{Name: queuePath, RateLimits: &taskspb.RateLimits{MaxDispatchesPerSecond: config.CONVERSATION_RATE_PER_SECOND()}},
-			}
-			_, err = client.CreateQueue(ctx, &createQueueRequest)
-			if err != nil {
-				return err
-			}
-			fmt.Println("queue created" + queuePath)
-		} else {
-			return err
-		}
+		return nil, err
 	}
 
 	// 送信するメッセージを作成
 	numTokens := 0
-	messages := []Message{}
+	messages := []ChatGPTAPIMessage{}
 
 	// systemメッセージを最初に載せる
-	messages = append(messages, Message{
+	messages = append(messages, ChatGPTAPIMessage{
 		Role:    "system",
-		Message: logs[0].Message,
+		Content: logs[0].Message,
 	})
 	if t, err := calcToken(logs[0].Message); err != nil {
-		return err
+		return nil, err
 	} else {
 		numTokens += t
 	}
 
-	insert := func(a []Message, index int, value Message) []Message {
+	insert := func(a []ChatGPTAPIMessage, index int, value ChatGPTAPIMessage) []ChatGPTAPIMessage {
 		if len(a) == index { // nil or empty slice or after last element
 			return append(a, value)
 		}
@@ -123,55 +105,75 @@ func (ai *aiServiceChatGPT3_5TurboImpl) SendRequest(ctx context.Context, convers
 	for i := len(logs) - 1; i != 0; i-- { // systemメッセージは飛ばす
 		t, err := calcToken(logs[i].Message)
 		if err != nil {
-			return err
+			return nil, err
 		}
 		if numTokens+t > tokenLimit {
 			// トークン数超過したらbreak
 			break
 		}
 		numTokens += t
-		messages = insert(messages, 1, Message{
+		messages = insert(messages, 1, ChatGPTAPIMessage{
 			Role:    logs[i].Role.String(),
-			Message: logs[i].Message,
+			Content: logs[i].Message,
 		})
 	}
 
 	fmt.Println(numTokens)
 
 	for _, message := range messages {
-		fmt.Println(message.Role + " " + message.Message)
+		fmt.Println(message.Role + " " + message.Content)
 	}
 
-	request := Request{
-		CallBackUrl: config.SELF_HOST() + "/conversations/" + conversationID.ToString() + "/reply",
-		AIModel:     modelName,
+	// ChatGPT APIに投げるリクエストを組み立てる
+	chatGPTAPIReq := ChatGPTAPIRequest{
+		Model:       modelName,
 		Temperature: temperature,
 		Messages:    messages,
 	}
 
-	body, err := json.Marshal(request)
+	jsonStr, err := json.Marshal(chatGPTAPIReq)
 	if err != nil {
-		return err
+		return nil, err
+	}
+	chatGPTAPIResp, err := reqChatGPTAPI(jsonStr)
+	if err == nil {
+		return ai_model.NewResponse(chatGPTAPIResp.Choices[0].Message.Content, ""), nil
+	} else {
+		fmt.Fprintf(os.Stderr, "ChatGPT API request error: %v\n", err)
+		return ai_model.NewResponse("", "ChatGPT API request error: "+err.Error()), nil
+	}
+}
+
+func reqChatGPTAPI(jsonStr []byte) (*ChatGPTAPIResponse, error) {
+
+	var chatGPTAPIResp ChatGPTAPIResponse
+
+	client := resty.New()
+
+	client.AddRetryCondition(
+		func(r *resty.Response, err error) bool {
+			return r.StatusCode() == http.StatusTooManyRequests
+		},
+	)
+	client.
+		SetRetryCount(3).
+		SetRetryWaitTime(5 * time.Second).
+		SetRetryMaxWaitTime(20 * time.Second)
+
+	resp, err := client.R().
+		SetHeader("Content-Type", "application/json").
+		SetHeader("Authorization", "Bearer "+config.CHATGPT_API_KEY()).
+		SetBody(jsonStr).
+		SetResult(&chatGPTAPIResp). // or SetResult(AuthSuccess{}).
+		Post("https://api.openai.com/v1/chat/completions")
+	if err != nil {
+		return nil, err
 	}
 
-	createTaskRequest := taskspb.CreateTaskRequest{
-		Parent: queuePath,
-		Task: &taskspb.Task{
-			// https://godoc.org/google.golang.org/genproto/googleapis/cloud/tasks/v2#HttpRequest
-			MessageType: &taskspb.Task_HttpRequest{
-				HttpRequest: &taskspb.HttpRequest{
-					HttpMethod: taskspb.HttpMethod_POST,
-					Url:        config.REQUESTOR_HOST() + "/openai_chat_gpt",
-					Body:       body,
-				},
-			},
-		}}
-	_, err = client.CreateTask(context.Background(), &createTaskRequest)
-	if err != nil {
-		return err
+	if resp.StatusCode() != http.StatusOK {
+		return nil, fmt.Errorf("ChatGPT API request error: status code %d, body %s", resp.StatusCode(), string(resp.Body()))
 	}
-
-	return nil
+	return &chatGPTAPIResp, nil
 }
 
 func calcToken(text string) (int, error) {
